refactors timers, rng, damage

+++ src/player.rs
@@ -3,15 +3,15 @@ use bevy::prelude::*;
 
 #[derive(Component)]
 pub struct Player {
-    pub health: i32,
+    pub health: f32,
 }
 
 impl Player {
     pub fn new() -> Self {
-        Self { health: 100 }
+        Self { health: 100. }
     }
 
-    pub fn recieve_damage(&mut self, damage: i32) {
+    pub fn recieve_damage(&mut self, damage: f32) {
         self.health -= damage;
         println!("Player health: {}", self.health)
     }
@@ -44,7 +44,7 @@ fn kill_player(
     mut player_query: Query<(Entity, &Player)>,
 ) {
     let Ok((entity, player)) = player_query.get_single_mut() else { return; };
-    if player.health <= 0 {
+    if player.health <= 0. {
         commands.entity(entity).despawn();
     }
 }
